moved topographical layer code to function that produces the image
